+++ packages/babel-plugin/src/utils/traversers/types.tsx
@@ -1,0 +1,7 @@
+import type { NodePath } from '@babel/traverse';
+import type * as t from '@babel/types';
+
+export type Result<T> = {
+  node: t.Node;
+  path: NodePath<T>;
+};