+++ packages/codemods/src/plugins/types.ts
@@ -7,6 +7,10 @@ import type {
   JSXSpreadAttribute,
   Options,
   Program,
+  VariableDeclaration,
+  ASTPath,
+  TaggedTemplateExpression,
+  ArrowFunctionExpression,
 } from 'jscodeshift';
 
 // We want to ensure the config contract is correct so devs can get type safety
@@ -40,6 +44,20 @@ export type BuildRefAttributesContext<T> = ValidateConfig<
   }
 >;
 
+export type BuildAttributesContext<T> = ValidateConfig<
+  T,
+  {
+    // The original component declaration
+    originalNode: ASTPath<VariableDeclaration | TaggedTemplateExpression>;
+    // The existing component declaration
+    currentNode: ASTPath<VariableDeclaration | TaggedTemplateExpression>;
+    // The original node after transforms
+    transformedNode: VariableDeclaration | TaggedTemplateExpression | ArrowFunctionExpression;
+    // The composed node that's been created during transformation
+    composedNode: VariableDeclaration | TaggedTemplateExpression | null;
+  }
+>;
+
 /**
  * Interface for codemods that handle migration from CSS-in-JS libraries to Compiled
  */
@@ -51,6 +69,16 @@ export interface Transform {
    * @returns {ImportDeclaration} The import to replace config.currentNode
    */
   buildImport?<T>(context: BuildImportContext<T>): ImportDeclaration;
+
+  /**
+   * Build the compiled import replacing the existing import
+   *
+   * @param context {BuildImportContext} The context applied to the build import
+   * @returns {VariableDeclaration | TaggedTemplateExpression} The import to replace config.currentNode
+   */
+  buildAttributes?<T>(
+    context: BuildAttributesContext<T>
+  ): ASTPath<VariableDeclaration | TaggedTemplateExpression>;
 
   /**
    * Build the compiled ref attribute replacing innerRef attributes